backend/ocr/requirements.txt
# PaddleOCR and dependencies
paddlepaddle==2.5.2
paddleocr==2.7.3

# Web framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Image processing
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3

# HTTP client
httpx==0.25.2
